requirements.txt
fastapi==0.115.6
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.30
langchain-openai==0.3.0
openai==1.59.8
openpyxl==3.1.5
pypdf==5.1.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
tiktoken==0.8.0
uvicorn==0.34.0
chromadb==0.6.3